fixed the lightbox being pushed down, deleted the previous secondSection vue file
