--- components/partials/sections/home/SecondSection2.vue
@@ -1,151 +1,0 @@
-<template>
-    <section id="home_second_section" ref="section" class="bg-[#EDF1F8] py-16 lg:py-20"  >
-        <div class="container px-5 overflow-hidden">
-            <h1 class="w-full">Perché Usalo! è unico</h1>
-            <div class="mt-[2rem] flex flex-col md:flex-row md:items-center  md:gap-x-10 lg:mt-20 xl:gap-36">
-                <!-- gears -->
-                <div class="w-full flex items-center justify-center md:order-2 md:w-fit">
-                    <div class="translate-x-1.5 lg:translate-x-2.5">
-                        <!-- small gear -->
-                        <svg  class="second-section-small-gear lg:hidden"  width="94" height="98" viewBox="0 0 94 98" fill="none" xmlns="http://www.w3.org/2000/svg">
-                            <path d="M66.8526 11.1918L64.6737 13.5193C63.8523 14.3917 62.7991 15.0119 61.6376 15.3073C60.476 15.6026 59.2543 15.5609 58.1157 15.187L58.0392 15.1625C56.8943 14.806 55.8749 14.1308 55.1004 13.2159C54.3259 12.301 53.8284 11.1844 53.6662 9.99704L53.2491 6.83886C52.2839 -0.450404 41.715 -0.407455 40.8139 6.89126L40.4211 10.0514C40.2784 11.2291 39.8026 12.3417 39.0496 13.2588C38.2966 14.1758 37.2976 14.8593 36.1698 15.2291L35.9587 15.309C34.8302 15.6834 33.6185 15.729 32.4651 15.4405C31.3116 15.1519 30.2642 14.5413 29.4453 13.6798L27.2366 11.3638C22.166 6.03406 13.6437 12.28 17.2064 18.7112L18.7439 21.4907C19.3229 22.54 19.591 23.7324 19.5169 24.9284C19.4427 26.1244 19.0294 27.2746 18.3252 28.2444L18.2727 28.3161C17.58 29.2944 16.6219 30.0544 15.5114 30.5066C14.4009 30.9588 13.1841 31.0843 12.0046 30.8685L8.87698 30.2889C1.64471 28.9562 -1.5774 39.013 5.08194 42.1339L7.96468 43.4776C9.04636 43.9887 9.96012 44.7969 10.5992 45.8078C11.2382 46.8187 11.5761 47.9905 11.5734 49.1863L11.5733 49.2789C11.5875 50.4775 11.2599 51.6553 10.6287 52.6746C9.99756 53.6939 9.08899 54.5125 8.00931 55.0346L5.14354 56.4009C-1.49564 59.5717 1.81153 69.603 9.03923 68.2002L12.182 67.6008C13.3472 67.3769 14.5521 67.4883 15.6564 67.9223C16.7607 68.3562 17.7189 69.0947 18.4195 70.0518L18.5511 70.2329C19.2539 71.1888 19.6696 72.3251 19.7495 73.5085C19.8294 74.692 19.5703 75.8738 19.0025 76.9154L17.4844 79.7197C13.9771 86.1803 22.5542 92.354 27.5764 86.9806L29.7592 84.6544C30.5813 83.7819 31.6352 83.1618 32.7974 82.8666C33.9596 82.5715 35.1819 82.6136 36.321 82.988L36.3975 83.0124C37.5417 83.369 38.5605 84.0442 39.3343 84.9589C40.1082 85.8736 40.605 86.9898 40.7666 88.1767L41.1876 91.3361C42.1451 98.6229 52.7 98.5712 53.6112 91.28L54.0015 88.1148C54.1462 86.9373 54.6231 85.8254 55.3766 84.9087C56.1301 83.9921 57.129 83.3087 58.2567 82.9383L58.4691 82.8672C59.597 82.4933 60.8081 82.4481 61.9608 82.7371C63.1134 83.0261 64.1598 83.6373 64.9774 84.4991L67.1874 86.8113C72.2415 92.1399 80.7867 85.9013 77.2061 79.4602L75.6674 76.6845C75.0911 75.633 74.8265 74.4393 74.9045 73.2429C74.9824 72.0466 75.3998 70.8972 76.1076 69.9293L76.1613 69.8538C76.8556 68.8782 77.8132 68.12 78.9224 67.6679C80.0317 67.2157 81.2467 67.0882 82.4257 67.3002L85.5494 67.8785C92.7855 69.2124 96.0076 59.1557 89.3445 56.0336L86.4656 54.6911C85.3817 54.1804 84.466 53.3716 83.8257 52.3594C83.1854 51.3472 82.8471 50.1736 82.8504 48.9761L82.8493 48.8873C82.8357 47.6885 83.164 46.5106 83.7958 45.4915C84.4277 44.4724 85.3369 43.6543 86.4171 43.1328L89.2841 41.7627C95.9221 38.5957 92.6111 28.5632 85.3896 29.9596L82.2456 30.5628C81.0816 30.789 79.8772 30.68 78.7729 30.2483C77.6686 29.8166 76.7098 29.08 76.0083 28.1245L75.8767 27.9434C75.1735 26.9875 74.7573 25.8513 74.6767 24.6677C74.5961 23.484 74.8544 22.3019 75.4214 21.2596L76.9422 18.4604C80.4417 11.9847 71.8786 5.81971 66.8526 11.1918Z" stroke="#041E42" stroke-miterlimit="10"/>
-                        </svg>
-                        <!-- small gear lg-->
-                        <svg  class="second-section-small-gear hidden lg:block"  width="216" height="223" viewBox="0 0 216 223" fill="none" xmlns="http://www.w3.org/2000/svg">
-                            <mask id="path-1-inside-1_245_1021" fill="white">
-                            <path d="M177.937 158.251C176.697 160.11 175.385 161.916 174.002 163.668C173.228 164.65 172.729 165.82 172.556 167.058C172.383 168.296 172.542 169.558 173.017 170.715L180.751 189.459C181.323 190.861 181.432 192.41 181.061 193.879C180.69 195.348 179.858 196.659 178.689 197.619C175.271 200.442 171.69 203.059 167.965 205.459C166.692 206.282 165.187 206.674 163.675 206.574C162.162 206.475 160.722 205.891 159.567 204.908L144.178 191.758C143.226 190.947 142.076 190.404 140.846 190.186C139.616 189.968 138.35 190.081 137.178 190.516C135.091 191.291 132.964 191.982 130.798 192.59C129.594 192.928 128.503 193.581 127.635 194.482C126.767 195.383 126.154 196.499 125.86 197.716L121.125 217.416C120.766 218.889 119.944 220.207 118.781 221.177C117.617 222.147 116.173 222.717 114.661 222.802C112.463 222.934 110.25 223 108.022 223C105.794 223 103.578 222.934 101.374 222.802C99.8611 222.721 98.415 222.153 97.2504 221.182C96.0859 220.212 95.2652 218.891 94.9104 217.416L90.1398 197.716C89.846 196.5 89.2341 195.384 88.3671 194.483C87.5001 193.582 86.4092 192.929 85.2065 192.59C83.0432 191.982 80.9166 191.291 78.8266 190.516C77.6542 190.081 76.3873 189.966 75.1562 190.185C73.9251 190.403 72.7743 190.946 71.8223 191.758L56.4332 204.908C55.2772 205.891 53.8357 206.475 52.3222 206.574C50.8088 206.672 49.304 206.28 48.0308 205.454C44.3024 203.062 40.7213 200.447 37.3067 197.624C36.1329 196.667 35.2984 195.357 34.9269 193.888C34.5554 192.418 34.6668 190.868 35.2446 189.467L42.9743 170.724C43.4491 169.567 43.6081 168.305 43.435 167.067C43.262 165.828 42.763 164.659 41.9894 163.677C40.6029 161.916 39.2839 160.11 38.0322 158.26C37.3353 157.226 36.3766 156.395 35.2541 155.854C34.1316 155.312 32.8856 155.079 31.6435 155.177L11.4178 156.75C9.9103 156.859 8.40777 156.479 7.13277 155.666C5.85777 154.852 4.8781 153.65 4.33884 152.235C2.71879 148.102 1.34744 143.876 0.232144 139.578C-0.152518 138.112 -0.0572613 136.561 0.503903 135.153C1.06507 133.745 2.06252 132.554 3.34954 131.757L20.6293 121.129C21.6975 120.469 22.5707 119.535 23.1588 118.425C23.7469 117.314 24.0288 116.066 23.9754 114.81C23.9314 113.713 23.9094 112.608 23.9094 111.502C23.9094 110.397 23.9314 109.3 23.9754 108.182C24.0293 106.925 23.7476 105.677 23.1594 104.566C22.5713 103.455 21.6979 102.522 20.6293 101.862L3.35833 91.2391C2.06815 90.4465 1.06758 89.2582 0.505191 87.8507C-0.0571967 86.4432 -0.15141 84.8916 0.236541 83.4263C1.34904 79.1285 2.71896 74.9018 4.33884 70.769C4.88339 69.3565 5.86766 68.1574 7.14585 67.3491C8.42404 66.5408 9.92816 66.1663 11.4354 66.2812L31.6611 67.8403C32.9032 67.9391 34.1492 67.7058 35.2716 67.1641C36.3941 66.6225 37.3529 65.7919 38.0498 64.7574C39.2897 62.893 40.6088 61.0844 42.007 59.3316C42.7806 58.3502 43.2795 57.1803 43.4526 55.942C43.6257 54.7037 43.4667 53.4415 42.9919 52.2851L35.2402 33.5326C34.6657 32.1305 34.5559 30.5807 34.9272 29.1115C35.2985 27.6422 36.1311 26.3313 37.3023 25.3718C40.7236 22.5512 44.3089 19.9364 48.0396 17.5414C49.3128 16.7176 50.817 16.3264 52.3295 16.4255C53.842 16.5247 55.2824 17.1089 56.4376 18.0919L71.8267 31.2337C72.7779 32.0453 73.9277 32.5886 75.1579 32.8077C76.3882 33.0267 77.6545 32.9136 78.8266 32.48C80.9166 31.7049 83.0447 31.012 85.2109 30.4013C86.4127 30.0618 87.5028 29.4081 88.369 28.5072C89.2352 27.6063 89.8464 26.4907 90.1398 25.275L94.8796 5.58436C95.2344 4.10921 96.0551 2.78817 97.2197 1.81772C98.3842 0.847274 99.8304 0.279268 101.343 0.198182C103.542 0.0660597 105.758 0 107.991 0C110.225 0 112.438 0.0660597 114.63 0.198182C116.142 0.283424 117.586 0.852931 118.75 1.82271C119.913 2.7925 120.735 4.11121 121.094 5.58436L125.829 25.275C126.123 26.4926 126.735 27.6098 127.603 28.5116C128.471 29.4134 129.563 30.0672 130.767 30.4057C132.93 31.0134 135.057 31.7064 137.147 32.4844C138.32 32.9189 139.587 33.0324 140.818 32.8134C142.049 32.5943 143.2 32.0505 144.151 31.2381L159.54 18.0919C160.697 17.1109 162.137 16.5283 163.649 16.4299C165.161 16.3316 166.665 16.7228 167.938 17.5458C171.671 19.9397 175.258 22.5545 178.68 25.3762C179.851 26.3357 180.684 27.6466 181.055 29.1159C181.426 30.5851 181.317 32.1349 180.742 33.537L173.012 52.2719C172.538 53.4283 172.379 54.6905 172.552 55.9288C172.725 57.1671 173.224 58.337 173.997 59.3184C175.378 61.08 176.69 62.8872 177.933 64.7398C178.628 65.7789 179.587 66.6138 180.712 67.1587C181.836 67.7036 183.085 67.9388 184.33 67.8403L204.556 66.268C206.065 66.1502 207.572 66.5232 208.853 67.3317C210.134 68.1402 211.12 69.341 211.665 70.7558C213.287 74.8878 214.657 79.1147 215.768 83.413C216.153 84.878 216.057 86.428 215.496 87.8346C214.935 89.2413 213.937 90.4301 212.65 91.2259L195.357 101.866C194.287 102.525 193.411 103.459 192.823 104.571C192.235 105.683 191.954 106.933 192.011 108.19C192.055 109.287 192.077 110.392 192.077 111.502C192.077 112.612 192.077 113.704 192.011 114.801C191.954 116.059 192.233 117.31 192.822 118.423C193.41 119.536 194.286 120.47 195.357 121.129L212.637 131.761C213.926 132.557 214.924 133.747 215.486 135.156C216.047 136.564 216.141 138.116 215.755 139.583C214.639 143.88 213.268 148.107 211.648 152.24C211.102 153.654 210.115 154.854 208.835 155.661C207.554 156.469 206.047 156.841 204.538 156.723L184.312 155.151C183.071 155.056 181.827 155.294 180.706 155.838C179.586 156.383 178.631 157.216 177.937 158.251Z"/>
-                            </mask>
-                            <path d="M177.937 158.251C176.697 160.11 175.385 161.916 174.002 163.668C173.228 164.65 172.729 165.82 172.556 167.058C172.383 168.296 172.542 169.558 173.017 170.715L180.751 189.459C181.323 190.861 181.432 192.41 181.061 193.879C180.69 195.348 179.858 196.659 178.689 197.619C175.271 200.442 171.69 203.059 167.965 205.459C166.692 206.282 165.187 206.674 163.675 206.574C162.162 206.475 160.722 205.891 159.567 204.908L144.178 191.758C143.226 190.947 142.076 190.404 140.846 190.186C139.616 189.968 138.35 190.081 137.178 190.516C135.091 191.291 132.964 191.982 130.798 192.59C129.594 192.928 128.503 193.581 127.635 194.482C126.767 195.383 126.154 196.499 125.86 197.716L121.125 217.416C120.766 218.889 119.944 220.207 118.781 221.177C117.617 222.147 116.173 222.717 114.661 222.802C112.463 222.934 110.25 223 108.022 223C105.794 223 103.578 222.934 101.374 222.802C99.8611 222.721 98.415 222.153 97.2504 221.182C96.0859 220.212 95.2652 218.891 94.9104 217.416L90.1398 197.716C89.846 196.5 89.2341 195.384 88.3671 194.483C87.5001 193.582 86.4092 192.929 85.2065 192.59C83.0432 191.982 80.9166 191.291 78.8266 190.516C77.6542 190.081 76.3873 189.966 75.1562 190.185C73.9251 190.403 72.7743 190.946 71.8223 191.758L56.4332 204.908C55.2772 205.891 53.8357 206.475 52.3222 206.574C50.8088 206.672 49.304 206.28 48.0308 205.454C44.3024 203.062 40.7213 200.447 37.3067 197.624C36.1329 196.667 35.2984 195.357 34.9269 193.888C34.5554 192.418 34.6668 190.868 35.2446 189.467L42.9743 170.724C43.4491 169.567 43.6081 168.305 43.435 167.067C43.262 165.828 42.763 164.659 41.9894 163.677C40.6029 161.916 39.2839 160.11 38.0322 158.26C37.3353 157.226 36.3766 156.395 35.2541 155.854C34.1316 155.312 32.8856 155.079 31.6435 155.177L11.4178 156.75C9.9103 156.859 8.40777 156.479 7.13277 155.666C5.85777 154.852 4.8781 153.65 4.33884 152.235C2.71879 148.102 1.34744 143.876 0.232144 139.578C-0.152518 138.112 -0.0572613 136.561 0.503903 135.153C1.06507 133.745 2.06252 132.554 3.34954 131.757L20.6293 121.129C21.6975 120.469 22.5707 119.535 23.1588 118.425C23.7469 117.314 24.0288 116.066 23.9754 114.81C23.9314 113.713 23.9094 112.608 23.9094 111.502C23.9094 110.397 23.9314 109.3 23.9754 108.182C24.0293 106.925 23.7476 105.677 23.1594 104.566C22.5713 103.455 21.6979 102.522 20.6293 101.862L3.35833 91.2391C2.06815 90.4465 1.06758 89.2582 0.505191 87.8507C-0.0571967 86.4432 -0.15141 84.8916 0.236541 83.4263C1.34904 79.1285 2.71896 74.9018 4.33884 70.769C4.88339 69.3565 5.86766 68.1574 7.14585 67.3491C8.42404 66.5408 9.92816 66.1663 11.4354 66.2812L31.6611 67.8403C32.9032 67.9391 34.1492 67.7058 35.2716 67.1641C36.3941 66.6225 37.3529 65.7919 38.0498 64.7574C39.2897 62.893 40.6088 61.0844 42.007 59.3316C42.7806 58.3502 43.2795 57.1803 43.4526 55.942C43.6257 54.7037 43.4667 53.4415 42.9919 52.2851L35.2402 33.5326C34.6657 32.1305 34.5559 30.5807 34.9272 29.1115C35.2985 27.6422 36.1311 26.3313 37.3023 25.3718C40.7236 22.5512 44.3089 19.9364 48.0396 17.5414C49.3128 16.7176 50.817 16.3264 52.3295 16.4255C53.842 16.5247 55.2824 17.1089 56.4376 18.0919L71.8267 31.2337C72.7779 32.0453 73.9277 32.5886 75.1579 32.8077C76.3882 33.0267 77.6545 32.9136 78.8266 32.48C80.9166 31.7049 83.0447 31.012 85.2109 30.4013C86.4127 30.0618 87.5028 29.4081 88.369 28.5072C89.2352 27.6063 89.8464 26.4907 90.1398 25.275L94.8796 5.58436C95.2344 4.10921 96.0551 2.78817 97.2197 1.81772C98.3842 0.847274 99.8304 0.279268 101.343 0.198182C103.542 0.0660597 105.758 0 107.991 0C110.225 0 112.438 0.0660597 114.63 0.198182C116.142 0.283424 117.586 0.852931 118.75 1.82271C119.913 2.7925 120.735 4.11121 121.094 5.58436L125.829 25.275C126.123 26.4926 126.735 27.6098 127.603 28.5116C128.471 29.4134 129.563 30.0672 130.767 30.4057C132.93 31.0134 135.057 31.7064 137.147 32.4844C138.32 32.9189 139.587 33.0324 140.818 32.8134C142.049 32.5943 143.2 32.0505 144.151 31.2381L159.54 18.0919C160.697 17.1109 162.137 16.5283 163.649 16.4299C165.161 16.3316 166.665 16.7228 167.938 17.5458C171.671 19.9397 175.258 22.5545 178.68 25.3762C179.851 26.3357 180.684 27.6466 181.055 29.1159C181.426 30.5851 181.317 32.1349 180.742 33.537L173.012 52.2719C172.538 53.4283 172.379 54.6905 172.552 55.9288C172.725 57.1671 173.224 58.337 173.997 59.3184C175.378 61.08 176.69 62.8872 177.933 64.7398C178.628 65.7789 179.587 66.6138 180.712 67.1587C181.836 67.7036 183.085 67.9388 184.33 67.8403L204.556 66.268C206.065 66.1502 207.572 66.5232 208.853 67.3317C210.134 68.1402 211.12 69.341 211.665 70.7558C213.287 74.8878 214.657 79.1147 215.768 83.413C216.153 84.878 216.057 86.428 215.496 87.8346C214.935 89.2413 213.937 90.4301 212.65 91.2259L195.357 101.866C194.287 102.525 193.411 103.459 192.823 104.571C192.235 105.683 191.954 106.933 192.011 108.19C192.055 109.287 192.077 110.392 192.077 111.502C192.077 112.612 192.077 113.704 192.011 114.801C191.954 116.059 192.233 117.31 192.822 118.423C193.41 119.536 194.286 120.47 195.357 121.129L212.637 131.761C213.926 132.557 214.924 133.747 215.486 135.156C216.047 136.564 216.141 138.116 215.755 139.583C214.639 143.88 213.268 148.107 211.648 152.24C211.102 153.654 210.115 154.854 208.835 155.661C207.554 156.469 206.047 156.841 204.538 156.723L184.312 155.151C183.071 155.056 181.827 155.294 180.706 155.838C179.586 156.383 178.631 157.216 177.937 158.251Z" stroke="#041E42" stroke-width="2" mask="url(#path-1-inside-1_245_1021)"/>
-                        </svg>
-
-                    </div>
-                    
-                    <div class="flex justify-center items-center relative h-44 w-44 -translate-x-1.5 lg:h-[25.75rem] lg:w-[25.75rem] lg:-translate-x-2.5">
-                        <!-- big gear -->
-                        <svg   class="second-section-big-gear rotate-big w-full h-full absolute top-0 left-0 lg:hidden" w viewBox="0 0 179 178" fill="none" xmlns="http://www.w3.org/2000/svg">
-                            <path d="M142.97 152.729C142.378 151.583 142.169 150.277 142.375 149.004C142.582 147.73 143.192 146.557 144.116 145.656C144.701 145.09 145.293 144.505 145.891 143.901C146.794 142.976 147.97 142.366 149.247 142.16C150.523 141.955 151.832 142.166 152.979 142.761C159.118 145.923 164.915 137.967 160.019 133.1C159.1 132.191 158.498 131.011 158.301 129.734C158.105 128.457 158.325 127.151 158.929 126.008C159.322 125.272 159.7 124.534 160.056 123.781C160.63 122.626 161.56 121.687 162.709 121.1C163.858 120.514 165.164 120.312 166.437 120.523C173.268 121.638 176.321 112.281 170.148 109.159C168.994 108.581 168.057 107.647 167.474 106.495C166.891 105.344 166.694 104.036 166.912 102.764C167.057 101.947 167.188 101.125 167.303 100.299C167.489 99.0194 168.082 97.8338 168.995 96.9176C169.907 96.0014 171.091 95.4033 172.37 95.2118C179.208 94.1628 179.222 84.3161 172.383 83.2602C171.105 83.0658 169.924 82.4657 169.014 81.5486C168.103 80.6315 167.512 79.446 167.328 78.1674C167.211 77.3473 167.082 76.529 166.942 75.7124C166.725 74.4399 166.923 73.1313 167.507 71.98C168.091 70.8286 169.03 69.8954 170.186 69.3182C176.366 66.2073 173.32 56.8387 166.505 57.9443C165.229 58.1545 163.918 57.9484 162.768 57.3564C161.618 56.7644 160.69 55.8181 160.12 54.6575C159.762 53.901 159.382 53.168 158.991 52.4313C158.391 51.2871 158.176 49.9806 158.377 48.7047C158.577 47.4289 159.184 46.2515 160.106 45.3464C165.019 40.4847 159.239 32.5164 153.082 35.6695C151.936 36.2628 150.629 36.4721 149.354 36.2666C148.08 36.061 146.905 35.4515 146.003 34.5279C145.439 33.9388 144.854 33.3473 144.247 32.7535C143.321 31.8516 142.71 30.6759 142.505 29.4002C142.299 28.1245 142.51 26.8166 143.106 25.67C146.27 19.5353 138.309 13.742 133.439 18.6344C132.529 19.5533 131.349 20.1553 130.071 20.3515C128.793 20.5477 127.486 20.3277 126.343 19.7239C125.606 19.3315 124.868 18.9535 124.114 18.5982C122.958 18.0238 122.018 17.0937 121.432 15.945C120.845 14.7963 120.644 13.4897 120.857 12.2178C121.968 5.38339 112.6 2.33842 109.479 8.5038C108.9 9.65728 107.966 10.5943 106.813 11.1767C105.661 11.7591 104.352 11.9559 103.08 11.7382C102.261 11.594 101.436 11.4652 100.613 11.3476C99.3328 11.1625 98.1463 10.5704 97.2291 9.6589C96.3119 8.74741 95.7128 7.565 95.5204 6.28671C94.4684 -0.54367 84.6178 -0.560471 83.5589 6.2772C83.3639 7.55331 82.7635 8.73293 81.8463 9.64191C80.9291 10.5509 79.7438 11.141 78.4653 11.3252C77.6447 11.4427 76.8258 11.5714 76.0088 11.7114C74.7354 11.9283 73.426 11.7302 72.2739 11.1465C71.1218 10.5628 70.1881 9.62431 69.6105 8.46964C66.4977 2.29345 57.1231 5.33785 58.2295 12.1482C58.4398 13.4238 58.2335 14.7332 57.6412 15.8825C57.0488 17.0318 56.1019 17.9599 54.9405 18.5294C54.1835 18.8868 53.4478 19.2695 52.7129 19.6578C51.569 20.253 50.2642 20.4655 48.9904 20.2641C47.7166 20.0626 46.5412 19.4577 45.6372 18.5386C40.7724 13.6286 32.7991 19.4049 35.9542 25.5572C36.5474 26.7025 36.7568 28.0078 36.5516 29.281C36.3463 30.5541 35.7373 31.7277 34.8142 32.6288C34.2248 33.1924 33.6294 33.7914 33.0387 34.384C32.1363 35.3092 30.9598 35.9197 29.6833 36.125C28.4068 36.3303 27.0981 36.1195 25.9507 35.5239C19.8122 32.362 14.0153 40.3177 18.9108 45.1856C19.8302 46.0941 20.4326 47.2741 20.6289 48.5512C20.8253 49.8283 20.6051 51.1347 20.001 52.2771C19.6093 53.0145 19.2339 53.757 18.8745 54.5046C18.3001 55.6602 17.3693 56.6007 16.2193 57.1874C15.0693 57.774 13.7612 57.9758 12.4877 57.7629C5.64903 56.6524 2.60213 66.0138 8.7714 69.1335C9.92404 69.7143 10.8594 70.6498 11.4397 71.8021C12.02 72.9544 12.2144 74.2626 11.9942 75.5337C11.8476 76.3551 11.7211 77.1762 11.6034 77.9988C11.418 79.2753 10.827 80.4585 9.91769 81.3739C9.00838 82.2893 7.82882 82.8885 6.55291 83.0832C-0.281768 84.1345 -0.298582 93.9788 6.5434 95.0371C7.8203 95.232 9.00067 95.832 9.91022 96.7486C10.8198 97.6652 11.4103 98.8498 11.5946 100.128C11.7122 100.948 11.841 101.766 11.981 102.582C12.1986 103.856 12.0004 105.165 11.4159 106.317C10.8313 107.469 9.89129 108.403 8.73485 108.98C2.55714 112.088 5.60109 121.459 12.4181 120.351C13.693 120.146 14.9998 120.357 16.1457 120.951C17.2917 121.546 18.2161 122.493 18.7826 123.652C19.1439 124.401 19.5202 125.144 19.9118 125.878C20.5098 127.022 20.7241 128.327 20.5232 129.602C20.3223 130.877 19.7168 132.053 18.796 132.958C13.8828 137.819 19.6605 145.791 25.819 142.635C26.965 142.043 28.2709 141.834 29.5444 142.04C30.8178 142.246 31.9916 142.855 32.8927 143.777C33.459 144.363 34.0584 144.958 34.649 145.551C35.5748 146.453 36.1856 147.629 36.391 148.905C36.5965 150.18 36.3856 151.488 35.7896 152.635C32.6257 158.77 40.5865 164.563 45.4574 159.671C46.3664 158.752 47.5472 158.15 48.8251 157.953C50.1031 157.757 51.4102 157.977 52.5533 158.581C53.2896 158.975 54.0326 159.35 54.7823 159.707C55.9384 160.28 56.8796 161.21 57.467 162.358C58.0544 163.507 58.257 164.814 58.0449 166.086C56.9296 172.913 66.2922 175.964 69.4163 169.796C69.995 168.642 70.9294 167.705 72.0816 167.123C73.2338 166.54 74.5428 166.343 75.8154 166.561C76.6374 166.708 77.4566 166.837 78.2821 166.952C79.5623 167.137 80.7486 167.73 81.6654 168.642C82.5822 169.554 83.1807 170.737 83.3723 172.016C84.4219 178.85 94.2748 178.863 95.3314 172.029C95.5259 170.752 96.1264 169.571 97.0441 168.661C97.9618 167.752 99.148 167.161 100.427 166.977C101.248 166.86 102.067 166.731 102.884 166.591C104.158 166.374 105.468 166.572 106.621 167.156C107.774 167.74 108.708 168.68 109.285 169.835C112.395 176.009 121.773 172.967 120.663 166.154C120.456 164.874 120.668 163.561 121.268 162.411C121.868 161.261 122.823 160.336 123.992 159.773C124.749 159.415 125.483 159.036 126.22 158.644C127.364 158.047 128.67 157.832 129.946 158.033C131.221 158.234 132.398 158.839 133.304 159.759C138.155 164.66 146.129 158.884 142.97 152.729Z" stroke="#041E42" stroke-miterlimit="10"/>
-                        </svg>
-                        <!-- big gear lg-->
-                        <svg   class="second-section-big-gear hidden lg:block rotate-big w-full h-full absolute top-0 left-0 " width="412" height="412" viewBox="0 0 412 412" fill="none" xmlns="http://www.w3.org/2000/svg">
-                            <mask id="path-1-inside-1_245_1020" fill="white">
-                            <path d="M388.667 228.702C388.158 232.797 387.519 236.853 386.749 240.868C386.479 242.252 386.638 243.685 387.202 244.977C387.767 246.269 388.711 247.358 389.91 248.1L401.012 254.931C402.299 255.726 403.291 256.92 403.836 258.33C404.382 259.741 404.45 261.292 404.033 262.745C402.704 267.396 401.218 271.98 399.575 276.495C399.061 277.923 398.092 279.142 396.817 279.965C395.542 280.788 394.033 281.169 392.52 281.051L379.509 280.036C378.102 279.931 376.696 280.257 375.48 280.972C374.263 281.687 373.294 282.756 372.701 284.036C370.937 287.767 369.069 291.425 367.097 295.012C366.407 296.245 366.111 297.659 366.247 299.065C366.384 300.471 366.946 301.802 367.86 302.879L376.321 312.797C377.299 313.954 377.872 315.399 377.951 316.912C378.031 318.425 377.613 319.922 376.762 321.175C374.06 325.171 371.225 329.066 368.256 332.861C367.327 334.056 366.032 334.914 364.569 335.303C363.106 335.691 361.556 335.59 360.157 335.013L348.102 330.03C346.797 329.495 345.36 329.371 343.982 329.673C342.604 329.976 341.351 330.691 340.39 331.724C337.595 334.719 334.699 337.624 331.704 340.438C330.669 341.398 329.953 342.652 329.649 344.031C329.346 345.41 329.471 346.848 330.007 348.155L334.989 360.22C335.563 361.619 335.663 363.168 335.274 364.63C334.886 366.092 334.03 367.386 332.837 368.316C329.045 371.289 325.149 374.126 321.148 376.827C319.896 377.679 318.398 378.097 316.885 378.017C315.373 377.938 313.928 377.365 312.771 376.386L302.859 367.933C301.781 367.019 300.45 366.457 299.043 366.321C297.637 366.184 296.222 366.481 294.989 367.17C291.408 369.154 287.753 371.021 284.023 372.77C282.744 373.364 281.676 374.334 280.962 375.55C280.248 376.767 279.922 378.172 280.028 379.579L281.042 392.592C281.161 394.105 280.78 395.615 279.957 396.89C279.134 398.165 277.915 399.134 276.488 399.648C271.976 401.289 267.396 402.775 262.749 404.106C261.295 404.526 259.743 404.458 258.331 403.913C256.919 403.367 255.725 402.374 254.931 401.086L248.097 389.955C247.363 388.745 246.277 387.788 244.983 387.213C243.69 386.638 242.251 386.473 240.861 386.741C236.849 387.514 232.792 388.153 228.692 388.659C227.29 388.826 225.972 389.418 224.915 390.355C223.859 391.292 223.113 392.53 222.779 393.902L219.728 406.585C219.37 408.057 218.54 409.371 217.365 410.327C216.19 411.283 214.734 411.827 213.22 411.876C210.821 411.959 208.411 412 205.989 412C203.567 412 201.156 411.959 198.758 411.876C197.244 411.828 195.787 411.284 194.612 410.328C193.437 409.372 192.607 408.057 192.25 406.585L189.199 393.924C188.864 392.552 188.119 391.314 187.062 390.376C186.006 389.438 184.688 388.845 183.286 388.676C179.188 388.174 175.135 387.536 171.125 386.763C169.741 386.49 168.306 386.647 167.013 387.212C165.721 387.777 164.631 388.724 163.89 389.924L157.056 401.055C156.261 402.341 155.068 403.332 153.658 403.876C152.249 404.421 150.699 404.489 149.247 404.071C144.594 402.745 140.014 401.259 135.508 399.613C134.081 399.098 132.862 398.129 132.039 396.855C131.216 395.58 130.834 394.07 130.953 392.557L131.967 379.548C132.074 378.141 131.749 376.734 131.035 375.517C130.321 374.299 129.253 373.329 127.973 372.735C124.242 370.989 120.583 369.12 116.994 367.13C115.759 366.441 114.343 366.147 112.936 366.287C111.529 366.426 110.199 366.993 109.123 367.911L99.207 376.36C98.0514 377.34 96.606 377.914 95.0926 377.994C93.5793 378.074 92.0816 377.655 90.8295 376.801C86.8318 374.102 82.9356 371.267 79.1407 368.294C77.9471 367.364 77.0906 366.069 76.7021 364.606C76.3135 363.143 76.4143 361.594 76.989 360.193L81.9714 348.128C82.502 346.826 82.6242 345.393 82.3219 344.019C82.0195 342.646 81.3068 341.397 80.2783 340.438C77.283 337.636 74.3832 334.737 71.5789 331.741C70.6187 330.708 69.3661 329.991 67.9882 329.688C66.6104 329.384 65.1727 329.509 63.8672 330.044L51.8037 335.013C50.404 335.59 48.8541 335.691 47.3912 335.303C45.9283 334.914 44.6332 334.056 43.704 332.861C40.7351 329.069 37.9 325.174 35.1986 321.175C34.3457 319.923 33.927 318.425 34.0066 316.912C34.0863 315.398 34.6599 313.953 35.6396 312.797L44.092 302.879C45.0047 301.803 45.5676 300.473 45.7054 299.068C45.8433 297.664 45.5496 296.25 44.8636 295.016C42.8795 291.433 41.0129 287.773 39.2639 284.036C38.6723 282.753 37.7029 281.681 36.4855 280.964C35.2681 280.247 33.8607 279.918 32.4517 280.023L19.4402 281.037C17.9277 281.155 16.4184 280.774 15.1439 279.951C13.8695 279.128 12.9003 277.909 12.3854 276.482C10.7452 271.975 9.25931 267.396 7.92773 262.745C7.50953 261.291 7.57831 259.74 8.12361 258.328C8.6689 256.917 9.66079 255.722 10.948 254.927L22.0813 248.087C23.2803 247.346 24.2251 246.256 24.7899 244.964C25.3546 243.672 25.5126 242.239 25.2427 240.855C24.4725 236.842 23.8332 232.787 23.3247 228.688C23.157 227.287 22.565 225.969 21.6281 224.913C20.6911 223.857 19.4539 223.113 18.0821 222.779L5.4189 219.728C3.94701 219.37 2.6328 218.54 1.67728 217.365C0.721758 216.189 0.177534 214.733 0.127857 213.219C0.0426126 210.814 0 208.403 0 205.987C0 203.57 0.0426126 201.158 0.127857 198.75C0.177534 197.236 0.721758 195.78 1.67728 194.605C2.6328 193.429 3.94701 192.599 5.4189 192.241L18.0821 189.19C19.4536 188.856 20.6905 188.111 21.6274 187.055C22.5642 185.999 23.1564 184.682 23.3247 183.281C23.8302 179.165 24.4696 175.106 25.2427 171.105C25.5126 169.721 25.3546 168.288 24.7899 166.996C24.2251 165.704 23.2803 164.615 22.0813 163.873L10.948 157.033C9.66079 156.238 8.6689 155.044 8.12361 153.632C7.57831 152.22 7.50953 150.669 7.92773 149.215C9.25343 144.567 10.7393 139.988 12.3854 135.478C12.9003 134.051 13.8695 132.832 15.1439 132.01C16.4184 131.187 17.9277 130.805 19.4402 130.923L32.4561 131.937C33.862 132.043 35.2669 131.718 36.4833 131.005C37.6997 130.292 38.6697 129.225 39.2639 127.946C41.0276 124.216 42.8942 120.557 44.8636 116.97C45.5518 115.737 45.8475 114.323 45.7112 112.918C45.5749 111.512 45.0131 110.181 44.1008 109.103L35.6484 99.2208C34.6701 98.064 34.0974 96.6187 34.0177 95.1057C33.9381 93.5927 34.356 92.0952 35.2074 90.8421C37.9118 86.8468 40.7469 82.9515 43.7128 79.1561C44.6386 77.9592 45.9309 77.0986 47.3921 76.706C48.8533 76.3134 50.4028 76.4104 51.8037 76.9821L63.8672 81.9652C65.1736 82.4992 66.6118 82.6219 67.9897 82.317C69.3676 82.0121 70.6198 81.294 71.5789 80.2586C74.3773 77.2599 77.2756 74.3612 80.2739 71.5624C81.3087 70.6028 82.0257 69.35 82.3291 67.9717C82.6325 66.5933 82.5077 65.1552 81.9714 63.8497L76.989 51.7844C76.4172 50.3854 76.318 48.8378 76.7064 47.3771C77.0948 45.9164 77.9496 44.6226 79.1407 43.6924C82.9326 40.729 86.8289 37.8935 90.8295 35.1859C92.0816 34.3322 93.5793 33.913 95.0926 33.9926C96.606 34.0723 98.0514 34.6464 99.207 35.6269L109.123 44.0981C110.199 45.012 111.529 45.5752 112.934 45.7124C114.339 45.8495 115.752 45.554 116.985 44.8654C120.562 42.884 124.222 41.0113 127.964 39.2473C129.243 38.6537 130.311 37.6838 131.025 36.4672C131.739 35.2506 132.064 33.8452 131.959 32.4386L130.944 19.4032C130.826 17.8906 131.208 16.381 132.031 15.1064C132.854 13.8317 134.072 12.8625 135.499 12.3475C140.011 10.7276 144.594 9.25767 149.247 7.93766C150.701 7.51714 152.253 7.58484 153.665 8.13041C155.077 8.67598 156.271 9.66934 157.064 10.9584L163.89 22.0491C164.631 23.25 165.721 24.1962 167.013 24.7612C168.306 25.3262 169.741 25.4831 171.125 25.2109C175.135 24.4436 179.187 23.8071 183.282 23.3015C184.682 23.1357 185.998 22.5473 187.055 21.6145C188.112 20.6817 188.86 19.4488 189.199 18.0802L192.25 5.42407C192.607 3.95163 193.437 2.63683 194.612 1.68105C195.787 0.725284 197.244 0.181252 198.758 0.132294C201.156 0.0440981 203.567 0 205.989 0C208.411 0 210.821 0.0411583 213.22 0.123475C214.734 0.173159 216.19 0.717457 217.365 1.67311C218.54 2.62877 219.37 3.94316 219.728 5.41525L222.779 18.0802C223.112 19.4538 223.858 20.6924 224.916 21.6297C225.974 22.567 227.293 23.1581 228.696 23.3235C232.797 23.8321 236.852 24.4701 240.861 25.2374C242.249 25.507 243.685 25.3461 244.978 24.7762C246.271 24.2064 247.36 23.2548 248.097 22.0491L254.922 10.9275C255.72 9.64403 256.917 8.65702 258.328 8.11716C259.74 7.57731 261.289 7.51426 262.74 7.93766C267.39 9.26649 271.973 10.7526 276.488 12.396C277.915 12.9103 279.134 13.8794 279.957 15.1542C280.78 16.429 281.161 17.9389 281.042 19.4517L280.028 32.465C279.928 33.8673 280.256 35.267 280.97 36.4783C281.683 37.6896 282.748 38.6553 284.023 39.2473C287.747 40.9936 291.403 42.8604 294.989 44.8478C296.222 45.5383 297.637 45.8353 299.043 45.6989C300.45 45.5626 301.781 44.9994 302.859 44.0849L312.771 35.6313C313.928 34.6528 315.373 34.08 316.885 34.0004C318.398 33.9208 319.896 34.3387 321.148 35.1903C325.149 37.8891 329.045 40.7261 332.837 43.7012C334.03 44.6312 334.886 45.9258 335.274 47.3876C335.663 48.8494 335.563 50.3981 334.989 51.7977L330.024 63.8541C329.488 65.1603 329.363 66.5991 329.667 67.9781C329.97 69.357 330.687 70.6106 331.722 71.5713C334.72 74.37 337.621 77.2687 340.425 80.2674C341.386 81.3013 342.638 82.0177 344.016 82.3211C345.394 82.6244 346.832 82.5003 348.137 81.9652L360.192 76.9821C361.592 76.4073 363.141 76.3065 364.604 76.6951C366.066 77.0837 367.362 77.9403 368.292 79.1341C371.26 82.9265 374.096 86.8219 376.797 90.8201C377.648 92.0734 378.065 93.5708 377.986 95.0836C377.906 96.5964 377.334 98.0416 376.356 99.1987L367.904 109.112C366.99 110.189 366.428 111.52 366.291 112.926C366.155 114.332 366.451 115.746 367.141 116.979C369.125 120.566 370.992 124.223 372.74 127.951C373.335 129.231 374.305 130.299 375.522 131.013C376.74 131.727 378.146 132.052 379.553 131.946L392.564 130.932C394.077 130.814 395.586 131.195 396.86 132.018C398.135 132.841 399.104 134.06 399.619 135.487C401.259 140 402.745 144.582 404.077 149.232C404.495 150.687 404.426 152.239 403.88 153.651C403.334 155.063 402.341 156.257 401.052 157.051L389.932 163.882C388.733 164.624 387.788 165.713 387.223 167.005C386.659 168.297 386.501 169.73 386.771 171.114C387.544 175.127 388.183 179.185 388.689 183.289C388.857 184.691 389.449 186.009 390.386 187.065C391.323 188.122 392.56 188.868 393.931 189.203L406.586 192.25C408.057 192.608 409.372 193.438 410.327 194.613C411.283 195.789 411.827 197.245 411.877 198.759C411.96 201.162 412 203.57 412 205.996C412 208.421 411.96 210.846 411.877 213.228C411.827 214.742 411.283 216.198 410.327 217.373C409.372 218.549 408.057 219.379 406.586 219.737L393.931 222.788C392.555 223.118 391.312 223.861 390.371 224.919C389.43 225.976 388.835 227.296 388.667 228.702Z"/>
-                            </mask>
-                            <path d="M388.667 228.702C388.158 232.797 387.519 236.853 386.749 240.868C386.479 242.252 386.638 243.685 387.202 244.977C387.767 246.269 388.711 247.358 389.91 248.1L401.012 254.931C402.299 255.726 403.291 256.92 403.836 258.33C404.382 259.741 404.45 261.292 404.033 262.745C402.704 267.396 401.218 271.98 399.575 276.495C399.061 277.923 398.092 279.142 396.817 279.965C395.542 280.788 394.033 281.169 392.52 281.051L379.509 280.036C378.102 279.931 376.696 280.257 375.48 280.972C374.263 281.687 373.294 282.756 372.701 284.036C370.937 287.767 369.069 291.425 367.097 295.012C366.407 296.245 366.111 297.659 366.247 299.065C366.384 300.471 366.946 301.802 367.86 302.879L376.321 312.797C377.299 313.954 377.872 315.399 377.951 316.912C378.031 318.425 377.613 319.922 376.762 321.175C374.06 325.171 371.225 329.066 368.256 332.861C367.327 334.056 366.032 334.914 364.569 335.303C363.106 335.691 361.556 335.59 360.157 335.013L348.102 330.03C346.797 329.495 345.36 329.371 343.982 329.673C342.604 329.976 341.351 330.691 340.39 331.724C337.595 334.719 334.699 337.624 331.704 340.438C330.669 341.398 329.953 342.652 329.649 344.031C329.346 345.41 329.471 346.848 330.007 348.155L334.989 360.22C335.563 361.619 335.663 363.168 335.274 364.63C334.886 366.092 334.03 367.386 332.837 368.316C329.045 371.289 325.149 374.126 321.148 376.827C319.896 377.679 318.398 378.097 316.885 378.017C315.373 377.938 313.928 377.365 312.771 376.386L302.859 367.933C301.781 367.019 300.45 366.457 299.043 366.321C297.637 366.184 296.222 366.481 294.989 367.17C291.408 369.154 287.753 371.021 284.023 372.77C282.744 373.364 281.676 374.334 280.962 375.55C280.248 376.767 279.922 378.172 280.028 379.579L281.042 392.592C281.161 394.105 280.78 395.615 279.957 396.89C279.134 398.165 277.915 399.134 276.488 399.648C271.976 401.289 267.396 402.775 262.749 404.106C261.295 404.526 259.743 404.458 258.331 403.913C256.919 403.367 255.725 402.374 254.931 401.086L248.097 389.955C247.363 388.745 246.277 387.788 244.983 387.213C243.69 386.638 242.251 386.473 240.861 386.741C236.849 387.514 232.792 388.153 228.692 388.659C227.29 388.826 225.972 389.418 224.915 390.355C223.859 391.292 223.113 392.53 222.779 393.902L219.728 406.585C219.37 408.057 218.54 409.371 217.365 410.327C216.19 411.283 214.734 411.827 213.22 411.876C210.821 411.959 208.411 412 205.989 412C203.567 412 201.156 411.959 198.758 411.876C197.244 411.828 195.787 411.284 194.612 410.328C193.437 409.372 192.607 408.057 192.25 406.585L189.199 393.924C188.864 392.552 188.119 391.314 187.062 390.376C186.006 389.438 184.688 388.845 183.286 388.676C179.188 388.174 175.135 387.536 171.125 386.763C169.741 386.49 168.306 386.647 167.013 387.212C165.721 387.777 164.631 388.724 163.89 389.924L157.056 401.055C156.261 402.341 155.068 403.332 153.658 403.876C152.249 404.421 150.699 404.489 149.247 404.071C144.594 402.745 140.014 401.259 135.508 399.613C134.081 399.098 132.862 398.129 132.039 396.855C131.216 395.58 130.834 394.07 130.953 392.557L131.967 379.548C132.074 378.141 131.749 376.734 131.035 375.517C130.321 374.299 129.253 373.329 127.973 372.735C124.242 370.989 120.583 369.12 116.994 367.13C115.759 366.441 114.343 366.147 112.936 366.287C111.529 366.426 110.199 366.993 109.123 367.911L99.207 376.36C98.0514 377.34 96.606 377.914 95.0926 377.994C93.5793 378.074 92.0816 377.655 90.8295 376.801C86.8318 374.102 82.9356 371.267 79.1407 368.294C77.9471 367.364 77.0906 366.069 76.7021 364.606C76.3135 363.143 76.4143 361.594 76.989 360.193L81.9714 348.128C82.502 346.826 82.6242 345.393 82.3219 344.019C82.0195 342.646 81.3068 341.397 80.2783 340.438C77.283 337.636 74.3832 334.737 71.5789 331.741C70.6187 330.708 69.3661 329.991 67.9882 329.688C66.6104 329.384 65.1727 329.509 63.8672 330.044L51.8037 335.013C50.404 335.59 48.8541 335.691 47.3912 335.303C45.9283 334.914 44.6332 334.056 43.704 332.861C40.7351 329.069 37.9 325.174 35.1986 321.175C34.3457 319.923 33.927 318.425 34.0066 316.912C34.0863 315.398 34.6599 313.953 35.6396 312.797L44.092 302.879C45.0047 301.803 45.5676 300.473 45.7054 299.068C45.8433 297.664 45.5496 296.25 44.8636 295.016C42.8795 291.433 41.0129 287.773 39.2639 284.036C38.6723 282.753 37.7029 281.681 36.4855 280.964C35.2681 280.247 33.8607 279.918 32.4517 280.023L19.4402 281.037C17.9277 281.155 16.4184 280.774 15.1439 279.951C13.8695 279.128 12.9003 277.909 12.3854 276.482C10.7452 271.975 9.25931 267.396 7.92773 262.745C7.50953 261.291 7.57831 259.74 8.12361 258.328C8.6689 256.917 9.66079 255.722 10.948 254.927L22.0813 248.087C23.2803 247.346 24.2251 246.256 24.7899 244.964C25.3546 243.672 25.5126 242.239 25.2427 240.855C24.4725 236.842 23.8332 232.787 23.3247 228.688C23.157 227.287 22.565 225.969 21.6281 224.913C20.6911 223.857 19.4539 223.113 18.0821 222.779L5.4189 219.728C3.94701 219.37 2.6328 218.54 1.67728 217.365C0.721758 216.189 0.177534 214.733 0.127857 213.219C0.0426126 210.814 0 208.403 0 205.987C0 203.57 0.0426126 201.158 0.127857 198.75C0.177534 197.236 0.721758 195.78 1.67728 194.605C2.6328 193.429 3.94701 192.599 5.4189 192.241L18.0821 189.19C19.4536 188.856 20.6905 188.111 21.6274 187.055C22.5642 185.999 23.1564 184.682 23.3247 183.281C23.8302 179.165 24.4696 175.106 25.2427 171.105C25.5126 169.721 25.3546 168.288 24.7899 166.996C24.2251 165.704 23.2803 164.615 22.0813 163.873L10.948 157.033C9.66079 156.238 8.6689 155.044 8.12361 153.632C7.57831 152.22 7.50953 150.669 7.92773 149.215C9.25343 144.567 10.7393 139.988 12.3854 135.478C12.9003 134.051 13.8695 132.832 15.1439 132.01C16.4184 131.187 17.9277 130.805 19.4402 130.923L32.4561 131.937C33.862 132.043 35.2669 131.718 36.4833 131.005C37.6997 130.292 38.6697 129.225 39.2639 127.946C41.0276 124.216 42.8942 120.557 44.8636 116.97C45.5518 115.737 45.8475 114.323 45.7112 112.918C45.5749 111.512 45.0131 110.181 44.1008 109.103L35.6484 99.2208C34.6701 98.064 34.0974 96.6187 34.0177 95.1057C33.9381 93.5927 34.356 92.0952 35.2074 90.8421C37.9118 86.8468 40.7469 82.9515 43.7128 79.1561C44.6386 77.9592 45.9309 77.0986 47.3921 76.706C48.8533 76.3134 50.4028 76.4104 51.8037 76.9821L63.8672 81.9652C65.1736 82.4992 66.6118 82.6219 67.9897 82.317C69.3676 82.0121 70.6198 81.294 71.5789 80.2586C74.3773 77.2599 77.2756 74.3612 80.2739 71.5624C81.3087 70.6028 82.0257 69.35 82.3291 67.9717C82.6325 66.5933 82.5077 65.1552 81.9714 63.8497L76.989 51.7844C76.4172 50.3854 76.318 48.8378 76.7064 47.3771C77.0948 45.9164 77.9496 44.6226 79.1407 43.6924C82.9326 40.729 86.8289 37.8935 90.8295 35.1859C92.0816 34.3322 93.5793 33.913 95.0926 33.9926C96.606 34.0723 98.0514 34.6464 99.207 35.6269L109.123 44.0981C110.199 45.012 111.529 45.5752 112.934 45.7124C114.339 45.8495 115.752 45.554 116.985 44.8654C120.562 42.884 124.222 41.0113 127.964 39.2473C129.243 38.6537 130.311 37.6838 131.025 36.4672C131.739 35.2506 132.064 33.8452 131.959 32.4386L130.944 19.4032C130.826 17.8906 131.208 16.381 132.031 15.1064C132.854 13.8317 134.072 12.8625 135.499 12.3475C140.011 10.7276 144.594 9.25767 149.247 7.93766C150.701 7.51714 152.253 7.58484 153.665 8.13041C155.077 8.67598 156.271 9.66934 157.064 10.9584L163.89 22.0491C164.631 23.25 165.721 24.1962 167.013 24.7612C168.306 25.3262 169.741 25.4831 171.125 25.2109C175.135 24.4436 179.187 23.8071 183.282 23.3015C184.682 23.1357 185.998 22.5473 187.055 21.6145C188.112 20.6817 188.86 19.4488 189.199 18.0802L192.25 5.42407C192.607 3.95163 193.437 2.63683 194.612 1.68105C195.787 0.725284 197.244 0.181252 198.758 0.132294C201.156 0.0440981 203.567 0 205.989 0C208.411 0 210.821 0.0411583 213.22 0.123475C214.734 0.173159 216.19 0.717457 217.365 1.67311C218.54 2.62877 219.37 3.94316 219.728 5.41525L222.779 18.0802C223.112 19.4538 223.858 20.6924 224.916 21.6297C225.974 22.567 227.293 23.1581 228.696 23.3235C232.797 23.8321 236.852 24.4701 240.861 25.2374C242.249 25.507 243.685 25.3461 244.978 24.7762C246.271 24.2064 247.36 23.2548 248.097 22.0491L254.922 10.9275C255.72 9.64403 256.917 8.65702 258.328 8.11716C259.74 7.57731 261.289 7.51426 262.74 7.93766C267.39 9.26649 271.973 10.7526 276.488 12.396C277.915 12.9103 279.134 13.8794 279.957 15.1542C280.78 16.429 281.161 17.9389 281.042 19.4517L280.028 32.465C279.928 33.8673 280.256 35.267 280.97 36.4783C281.683 37.6896 282.748 38.6553 284.023 39.2473C287.747 40.9936 291.403 42.8604 294.989 44.8478C296.222 45.5383 297.637 45.8353 299.043 45.6989C300.45 45.5626 301.781 44.9994 302.859 44.0849L312.771 35.6313C313.928 34.6528 315.373 34.08 316.885 34.0004C318.398 33.9208 319.896 34.3387 321.148 35.1903C325.149 37.8891 329.045 40.7261 332.837 43.7012C334.03 44.6312 334.886 45.9258 335.274 47.3876C335.663 48.8494 335.563 50.3981 334.989 51.7977L330.024 63.8541C329.488 65.1603 329.363 66.5991 329.667 67.9781C329.97 69.357 330.687 70.6106 331.722 71.5713C334.72 74.37 337.621 77.2687 340.425 80.2674C341.386 81.3013 342.638 82.0177 344.016 82.3211C345.394 82.6244 346.832 82.5003 348.137 81.9652L360.192 76.9821C361.592 76.4073 363.141 76.3065 364.604 76.6951C366.066 77.0837 367.362 77.9403 368.292 79.1341C371.26 82.9265 374.096 86.8219 376.797 90.8201C377.648 92.0734 378.065 93.5708 377.986 95.0836C377.906 96.5964 377.334 98.0416 376.356 99.1987L367.904 109.112C366.99 110.189 366.428 111.52 366.291 112.926C366.155 114.332 366.451 115.746 367.141 116.979C369.125 120.566 370.992 124.223 372.74 127.951C373.335 129.231 374.305 130.299 375.522 131.013C376.74 131.727 378.146 132.052 379.553 131.946L392.564 130.932C394.077 130.814 395.586 131.195 396.86 132.018C398.135 132.841 399.104 134.06 399.619 135.487C401.259 140 402.745 144.582 404.077 149.232C404.495 150.687 404.426 152.239 403.88 153.651C403.334 155.063 402.341 156.257 401.052 157.051L389.932 163.882C388.733 164.624 387.788 165.713 387.223 167.005C386.659 168.297 386.501 169.73 386.771 171.114C387.544 175.127 388.183 179.185 388.689 183.289C388.857 184.691 389.449 186.009 390.386 187.065C391.323 188.122 392.56 188.868 393.931 189.203L406.586 192.25C408.057 192.608 409.372 193.438 410.327 194.613C411.283 195.789 411.827 197.245 411.877 198.759C411.96 201.162 412 203.57 412 205.996C412 208.421 411.96 210.846 411.877 213.228C411.827 214.742 411.283 216.198 410.327 217.373C409.372 218.549 408.057 219.379 406.586 219.737L393.931 222.788C392.555 223.118 391.312 223.861 390.371 224.919C389.43 225.976 388.835 227.296 388.667 228.702Z" stroke="#041E42" stroke-width="2" mask="url(#path-1-inside-1_245_1020)"/>
-                        </svg>
-
-                        <transition name="gear-img">
-                            <img v-if="currentText===1" class="absolute top-1/2 -translate-y-1/2 left-1/2 -translate-x-1/2 lg:w-[9.4rem]" src="/images/home/second-section-img1.svg" alt="" srcset="">
-                            <img v-else-if="currentText===2" class="absolute top-1/2 -translate-y-1/2 left-1/2 -translate-x-1/2 lg:w-[7.8rem]" src="/images/home/second-section-img2.svg" alt="" srcset="">
-                            <img v-else-if="currentText===3" class="absolute top-1/2 -translate-y-1/2 left-1/2 -translate-x-1/2 lg:w-[9.2rem]" src="/images/home/second-section-img3.svg" alt="" srcset="">
-                        </transition>
-                        
-                    </div>
-                    
-                </div>
-                <!-- text content -->
-                <div class="w-full flex flex-col gap-y-4 font-light mt-10 md:order-1 md:max-w-[26.875rem] lg:text-lg">
-                    <div ref="text1" class="cursor-pointer">L'importo del <span class="font-bold">canone è collegato al valore</span> del bene: ogni anno pagherai meno del precedente!<br><br> </div>
-                    <div ref="text2" class="cursor-pointer"><span class="font-bold">Non c'è scadenza:</span> potrai usare il bene fino a che vorrai, con un contratto che si rinnova tacitamente ogni anno.<br><br> </div>
-                    <div ref="text3" class="cursor-pointer">In caso di uscita dal contratto, potrai beneficiare del possibile residuo valore del bene, presentandoci un acquirente!<br><br> </div>
-                </div>
-            </div>
-        </div>
-
-    </section>
-</template>
-
-<script setup>
-import { gsap } from "gsap";
-import { ScrollTrigger } from "gsap/ScrollTrigger";
-
-const currentText = ref(0)
-
-const section = ref(null)
-const text1 = ref(null)
-const text2 = ref(null)
-const text3 = ref(null)
-
-
-function initiateGsap() {
-    gsap.registerPlugin(ScrollTrigger)
-
-    const timeline = gsap.timeline(
-        {
-            scrollTrigger: {
-                trigger: section.value,
-                toggleActions:"restart none none none",
-                start: "top top",
-                end: "bottom top",
-                pinSpacing:false,
-                scrub: 1,
-                pin: false,
-            }
-        }
-    )
-    // first part
-    .from (text1.value, {opacity:0.5, y: innerHeight * 1.5,
-                        onStart:()=>{
-                            currentText.value = 1;
-                            text2.value.classList.remove('transition-opacity');
-                            text2.value.style.opacity = '1';
-                        }, 
-                        onReverseComplete:()=>{currentText.value = 0}},0)
-    .to (".second-section-big-gear", {rotation:36, duration:1.5},0)
-    .to (".second-section-small-gear", {rotation:-72, duration:1.5},0)
-    // second part
-    .from (text2.value, {opacity:0.5 ,y: innerHeight * 1.5,
-                        onStart:()=>{
-                            currentText.value = 2; 
-                            text2.value.classList.remove('transition-opacity');
-                            text2.value.style.opacity = '1';
-                            text1.value.classList.add('transition-opacity');
-                            text1.value.style.opacity = '0.5';
-                        }, 
-                        onReverseComplete:()=>{currentText.value = 1}},1)
-    .to (".second-section-big-gear", {rotation:72, duration:1.5},1)
-    .to (".second-section-small-gear", {rotation:-144, duration:1.5},1)
-    // third part
-    .from (text3.value, {opacity:0.5 ,y: innerHeight * 1.5,
-                        onStart:()=>{
-                            currentText.value = 3;
-                            text3.value.classList.remove('transition-opacity');
-                            text3.value.style.opacity = '1';
-                            text2.value.classList.add('transition-opacity');
-                            text2.value.style.opacity = '0.5';
-                        }, 
-                        onReverseComplete:()=>{currentText.value = 2}},2)
-    .to (".second-section-big-gear", {rotation:108, duration:1.5},2)
-    .to (".second-section-small-gear", {rotation:-216, duration:1.5},2)
-
-    .to (section.value, {rotate:0},3)
-}
-
-onMounted(() => {
-    initiateGsap()
-    
-  
-})
-
-
-</script>
-
-<style scoped>
-.sky-gradient{
-    background: radial-gradient(415.52% 159.44% at 3.79% -12.24%, #C9DDF2 0%, #E7F3FF 58.85%, #C9DDF2 100%);
-    box-shadow: 0px 24px 30px 0px rgba(0, 0, 0, 0.04), 0px 4px 14px 0px rgba(0, 0, 0, 0.05);
-    
-}
-.gear-img-enter-active {
-  transition: opacity 0.3s ease 0.3s;
-}
-.gear-img-leave-active {
-  transition: opacity 0.3s ease;
-}
-
-.gear-img-enter-from,
-.gear-img-leave-to {
-  opacity: 0;
-}
-
-
-</style>
